requirements.txt
PyTikTokAPI==0.0.5
PyYAML==5.4.1
requests==2.25.1
